Split lib/auth-form.ts into auth-errors.ts and use-shake.ts
auth-form.ts bundled two unrelated concerns: friendlyAuthError
(server error-message translation) and useShake (a button
shake-animation hook). fallow flagged it as the last remaining
refactoring target: 5 dependents split across the two exports, so
any change to one risked rippling through importers that only used
the other.

Split into auth-errors.ts and use-shake.ts, repointed all 5 importers
(auth.tsx, billing.tsx, reset-password.tsx/ts, tests), and renamed the
test file to match. No behavior change; live-verified a failed sign-in
(friendlyAuthError toast) plus reset-password and billing pages
loading cleanly with the new import paths.
